chore: shift hooks to common
