+++ src/components/common/hooks/constants.ts
@@ -1,0 +1,15 @@
+export const NO_FILE_SELECTED_STATUS = {
+    message: {
+        data: null,
+        description: 'No file selected',
+    },
+    status: false,
+}
+
+export const FILE_READING_FAILED_STATUS = {
+    message: {
+        data: null,
+        description: 'File reading failed',
+    },
+    status: false,
+}
