+++ src/components/common/index.ts
@@ -34,3 +34,4 @@ export * from './helpers/compareVersion'
 export * from './header/PageHeader'
 export * from './error/ErrorBar'
 export * from './DetectBottom'
+export * from './hooks'
