+++ src/components/scopedVariables/Descriptor.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import { TippyCustomized, TippyTheme } from '@devtron-labs/devtron-fe-common-lib'
 import SearchBar from './DescriptorSearchBar'
-import { validator } from './utils/helpers'
-import { DescriptorInterface, ReadFileAs } from './types'
+import { validator } from './utils'
+import { DescriptorInterface } from './types'
+import { ReadFileAs } from '../common/hooks/types'
 import { ReactComponent as ICHelpOutline } from '../../assets/img/ic-help-outline.svg'
 import { ReactComponent as QuestionFilled } from '../../assets/icons/ic-help.svg'
 import { ReactComponent as ICUpload } from '../../assets/icons/ic-upload-blue.svg'
