+++ src/components/scopedVariables/ScopedVariables.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react'
 import { ErrorScreenNotAuthorized, Progressing, Reload } from '@devtron-labs/devtron-fe-common-lib'
 import SavedVariablesView from './SavedVariables'
 import UploadScopedVariables from './UploadScopedVariables'
-import { sortVariables } from './utils/helpers'
-import { getScopedVariablesJSON } from './utils/service'
+import { sortVariables } from './utils'
+import { getScopedVariablesJSON } from './service'
 import { useAsync } from '../common'
 import { ScopedVariablesDataInterface, ScopedVariablesInterface } from './types'
 import './styles.scss'
