+++ src/components/scopedVariables/ScopedVariablesEditor.tsx
@@ -4,8 +4,8 @@ import Tippy from '@tippyjs/react'
 import Descriptor from './Descriptor'
 import CodeEditor from '../CodeEditor/CodeEditor'
 import { ButtonWithLoader } from '../common'
-import { parseYAMLStringToObj, parseIntoYAMLString, sortVariables } from './utils/helpers'
-import { postScopedVariables, getScopedVariablesJSON } from './utils/service'
+import { parseYAMLStringToObj, parseIntoYAMLString, sortVariables } from './utils'
+import { postScopedVariables, getScopedVariablesJSON } from './service'
 import { ScopedVariablesDataInterface, ScopedVariablesEditorInterface } from './types'
 import { ReactComponent as Close } from '../../assets/icons/ic-close.svg'
 import { ReactComponent as ICArrowRight } from '../../assets/icons/ic-arrow-right.svg'
