+++ src/components/scopedVariables/constants.tsx
@@ -11,20 +11,6 @@ export const DROPDOWN_ITEMS = ['Download saved file', 'Download template']
 export const TABLE_LIST_HEADINGS = ['VARIABLE NAMES', 'DESCRIPTION']
 
 // File Reader error messages
-export const FILE_READING_FAILED_STATUS = {
-    message: {
-        data: null,
-        description: 'File reading failed',
-    },
-    status: false,
-}
-export const NO_FILE_SELECTED_STATUS = {
-    message: {
-        data: null,
-        description: 'No file selected',
-    },
-    status: false,
-}
 export const PARSE_ERROR_STATUS = {
     message: {
         data: null,
